Reference

Reference

Agent generation

RSUOptimization.Agent — Type.

Agent type stores information about agents

Fields

smart : logical value indicating if agent can communicate with RSUs

start_node : starting point of agent's route

end_node : ending point of agent's route

route : array of nodes determining agent's route (may be changed by re-routing)

travel_time : time spend in simulation

edge : current edge agent is on

pos : position on current edge

active : indicates if agent is active in simulation

source

RSUOptimization.pick_random_node — Function.

pick_random_node function is used to set starting and ending node of agents. Nodes are randomly chosen from set of rectangles corresponding to areas on map.

Input parameters

OSMmap : mapData type object with road network data

rects : vector of tuples with two Latitude-Longitude point interpreted as a set of rectangle areas

source

RSUOptimization.generate_agents — Function.

generate_agents function creating vector of agents and returning travel time for initial routes travelled with maximal speed

Input parameters

OSMmap : MapData type object with road network data

N : number of agents to be generated

StartArea : vector of points corresponding to area from which agents randomly pick starting point

EndArea : vector of points corresponding to area from which agents randomly pick ending point

α : percentage of smart agentss

source

RSU location optimization

RSUOptimization.get_agent_coordinates — Function.

get_agent_coordinates return agents coordinates in ENU system

Input parameters

OSMmap : MapData type object with road network data

inAgent : agent which current coordinates are requested

source

Missing docstring.

Missing docstring for optimize_RSU_location. Check Documenter's build log for details.

Rerouting

RSUOptimization.k_shortest_path_rerouting! — Function.

k_shortest_path_rerouting! change agents current route according to traffic conditions

Input parameters

OSMmap : MapData type object with road network data

inAgent : rerouting agent

speeds : current speeds matrix

k : number of fastest routes returned

T : control variable for probability distribution

source

Vehicular communication

RSUOptimization.send_weights_update — Function.

send_weights_update function mark agents receiving weights update

Input parameters

Agents : set of agents created with generate_agents function

OSMmap : MapData type object with road network data

RSUs : vector with RSUs objects

range : range of RSUs transfer

source

Traffic model

RSUOptimization.get_max_densities — Function.

get_max_densities calculate maximal traffic density (number of agents) on each road in given network

Input parameters

OSMmap : MapData type object with road network data

density_factor : road length reserved for one vehicle

source

RSUOptimization.update_weights! — Function.

update_weights! change speeds in given speedmatrix for edges listed in newdensities

Input parameters

speed_matrix : matrix with average speeds on edges

new_densities : dictionary with edges as keys and new traffic density as value

V_max : matrix with maximal speeds on edges

V_min : minimal speed on road

source

RSUOptimization.traffic_constants — Function.

traffic_constants create maximal traffic densities and speeds matrices

Input parameters

OSMmap : MapData type object with road network data

density_factor : road length reserved for one vehicle

source

RSUOptimization.init_traffic_variables — Function.

init_traffic_variables create data used in calculating velocities change during simulation

Input parameters

OSMmap : MapData type object with road network data

Agents : set of agents created with generate_agents function

optimization_stats : switch for creating data used in RSUs optimization algorithm

source

RSUOptimization.next_edge — Function.

next_edge returns time for nearest edge switch and ID of agent performing the switch

Input parameters

Agents : set of agents created with generate_agents function

speeds : current speeds matrix

lengths : matrix with road lengths

source

RSUOptimization.update_event_agent! — Function.

update_event_agent! update densities matrix, progress agents to next edge and deactivate agents

Input parameters

inAgent : agent connected with occuring event

curr_time : current simulation time

densities : current traffic densitites matrix

vertices_map : mapping from nodes to vertices

debug : debug switch

source

Missing docstring.

Missing docstring for update_smart_densities!. Check Documenter's build log for details.

RSUOptimization.update_agents_position! — Function.

update_agents_position! change agents position on edge according to given time passed

Input parameters

Agents : set of agents created with generate_agents function

time_passed : time passed since last event

speeds : current speeds matrix

source