Reference

Reference

Agent generation

Missing docstring.

Missing docstring for Agent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pick_random_node. Check Documenter's build log for details.

Missing docstring.

Missing docstring for generate_agents. Check Documenter's build log for details.

RSU location optimization

Missing docstring.

Missing docstring for get_agent_coordinates. Check Documenter's build log for details.

Missing docstring.

Missing docstring for optimize_RSU_location. Check Documenter's build log for details.

Rerouting

Missing docstring.

Missing docstring for k_shortest_path_rerouting!. Check Documenter's build log for details.

Vehicular communication

Missing docstring.

Missing docstring for send_weights_update. Check Documenter's build log for details.

Traffic model

RSUOptimization.get_max_densities — Function.

get_max_densities calculate maximal traffic density (number of agents) on each road in given network

Input parameters

OSMmap : MapData type object with road network data

density_factor : road length reserved for one vehicle

source

RSUOptimization.update_weights! — Function.

update_weights! change speeds in given speedmatrix for edges listed in newdensities

Input parameters

speed_matrix : matrix with average speeds on edges

new_densities : dictionary with edges as keys and new traffic density as value

V_max : matrix with maximal speeds on edges

V_min : minimal speed on road

source

RSUOptimization.traffic_constants — Function.

traffic_constants create maximal traffic densities and speeds matrices

Input parameters

OSMmap : MapData type object with road network data

density_factor : road length reserved for one vehicle

source

Missing docstring.

Missing docstring for init_traffic_variables. Check Documenter's build log for details.

Missing docstring.

Missing docstring for next_edge. Check Documenter's build log for details.

Missing docstring.

Missing docstring for update_event_agent!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for update_smart_densities!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for update_agents_position!. Check Documenter's build log for details.