Reference

Reference

Agent generation

RSUOptimization.Rect — Type.

Rect type represent rectangle area on map

Fields

p1 : Latitude-Longitude coordinates of first rectangle corner

p2 : Latitude-Longitude coordinates of second rectangle corner

source

RSUOptimization.Agent — Type.

Agent type stores information about agents

Fields

smart : logical value indicating if agent can communicate with RSUs

start_node : starting point of agent's route

end_node : ending point of agent's route

route : array of nodes determining agent's route (may be changed by re-routing)

travel_time : time spend in simulation

edge : current edge agent is on

pos : position on current edge

active : indicates if agent is active in simulation

source

RSUOptimization.pick_random_node — Function.

pick_random_node function is used to set starting and ending node of agents. Nodes are randomly chosen from set of rectangles corresponding to areas on map.

Input parameters

OSMmap : mapData type object with road network data

rects : vector of Rect types interpreted as a set of rectangle areas

source

RSUOptimization.generate_agents — Function.

generate_agents function creating vector of agents and returning travel time for initial routes travelled with maximal speed

Input parameters

OSMmap : MapData type object with road network data

N : number of agents to be generated

StartArea : vector of points corresponding to area from which agents randomly pick starting point

EndArea : vector of points corresponding to area from which agents randomly pick ending point

α : percentage of smart agentss

source

RSU location optimization

RSUOptimization.RSU — Type.

RSU type stores information about set of Road Side units in one node

Fields

node : ID of node where RSU is located

ENU : ENU coordinates of the node

count : number of units in node

total_thput : total throughput of all units in the node

source

RSUOptimization.get_agent_coordinates — Function.

get_agent_coordinates return agents coordinates in ENU system

Input parameters

OSMmap : MapData type object with road network data

inAgent : agent which current coordinates are requested

source

RSUOptimization.calculate_RSU_location — Function.

calculate_RSU_location function returns dictionary with number of RSUs in given nodes.

Input parameters

OSMmap : MapData type from OpenStreetMapX package

range : range of RSUs

throughput : number of agents RSU can serve at once

div_coeff : adjustment factor for calculating number of RSUs in node

source

RSUOptimization.adjust_RSU_availability! — Function.

adjust_RSU_availability! function adjust RSUs location and number to meet service availability and utilization criteria

Input parameters

OSMmap : MapData type from OpenStreetMapX package

RSUs : vector with RSUs used in simulation

failed_coor : vector of vectors with agents coordinates missing an update

range : range of RSUs

throughput : number of agents RSU can serve at once

source

RSUOptimization.adjust_RSU_utilization! — Function.

adjust_RSU_utilization! function

Input parameters

RSUs : vector with RSUs used in simulation

RSUs_util : vector with RSUs percentage utilization in every weigths update

throughput : number of agents RSU can serve at once

source

RSUOptimization.gather_statistics — Function.

gather_statistics return various measures describing ITS model quality

Input parameters

smart_ind : vector with true flags for smart agents

times_base : agents travelling time in base scenario

times_ITS : agents travelling time in ITS scenario

srvc_avblty : vector with percentage service availability (from simulation_ITS)

RSUs_util : vector of dictionaries with RSUs utilization (from simulation_ITS)

RSUs : vector of RSUs

source

Rerouting

RSUOptimization.k_shortest_path_rerouting! — Function.

k_shortest_path_rerouting! change agents current route according to traffic conditions

Input parameters

OSMmap : MapData type object with road network data

inAgent : rerouting agent

speeds : current speeds matrix

k : number of fastest routes returned

T : control variable for probability distribution

source

Vehicular communication

RSUOptimization.send_weights_update — Function.

send_weights_update function mark agents receiving weights update

Input parameters

Agents : set of agents created with generate_agents function

OSMmap : MapData type object with road network data

RSUs : vector with RSUs objects

range : range of RSUs transfer

V2V : V2V hybrid model switch

V2V_range : range of V2V communication

V2V_throughput : throughput of V2V master to slaves communication

source

Traffic model

RSUOptimization.get_max_densities — Function.

get_max_densities calculate maximal traffic density (number of agents) on each road in given network

Input parameters

OSMmap : MapData type object with road network data

density_factor : road length reserved for one vehicle

source

RSUOptimization.update_weights! — Function.

update_weights! change speeds in given speedmatrix for edges listed in newdensities

Input parameters

speed_matrix : matrix with average speeds on edges

new_densities : dictionary with edges as keys and new traffic density as value

V_max : matrix with maximal speeds on edges

V_min : minimal speed on road

source

RSUOptimization.traffic_constants — Function.

traffic_constants create maximal traffic densities and speeds matrices

Input parameters

OSMmap : MapData type object with road network data

density_factor : road length reserved for one vehicle

source

RSUOptimization.init_traffic_variables — Function.

init_traffic_variables create data used in calculating velocities change during simulation

Input parameters

OSMmap : MapData type object with road network data

Agents : set of agents created with generate_agents function

optimization_stats : switch for creating data used in RSUs optimization algorithm

source

RSUOptimization.next_edge — Function.

next_edge returns time for nearest edge switch and ID of agent performing the switch

Input parameters

Agents : set of agents created with generate_agents function

speeds : current speeds matrix

lengths : matrix with road lengths

source

RSUOptimization.update_event_agent! — Function.

update_event_agent! update densities matrix, progress agents to next edge and deactivate agents

Input parameters

inAgent : agent connected with occuring event

curr_time : current simulation time

densities : current traffic densitites matrix

vertices_map : mapping from nodes to vertices

source

RSUOptimization.update_agents_position! — Function.

update_agents_position! change agents position on edge according to given time passed

Input parameters

Agents : set of agents created with generate_agents function

time_passed : time passed since last event

speeds : current speeds matrix

source